Issue #159/EP-3578 move `DataCube.execute_local_udf` to `openeo.udf` package
`execute_local_udf` had nothing to do with  `rest.DataCube`
