+++ tests/udf/test_run_code.py
@@ -5,11 +5,13 @@ import pandas
 import pytest
 import xarray
 
+from openeo.rest.conversions import datacube_to_file
 from openeo.udf.run_code import run_udf_code, _get_annotation_str, _annotation_is_pandas_series, \
-    _annotation_is_udf_datacube, \
-    _annotation_is_udf_data
+    _annotation_is_udf_datacube, _annotation_is_udf_data, execute_local_udf
 from openeo.udf.udf_data import UdfData
 from openeo.udf.xarraydatacube import XarrayDataCube
+from .test_xarraydatacube import _build_xdc
+from .. import as_path
 
 UDF_CODE_PATH = Path(__file__).parent / "udf_code"
 
@@ -150,3 +152,88 @@ def test_run_udf_code_statistics():
         "type": "dict",
         "description": "Statistical data sum, min, max and mean for each raster collection cube as dict"
     }
+
+
+def _ndvi(red, nir):
+    return (nir - red) / (nir + red)
+
+
+def test_execute_local_udf_basic():
+    xdc = _build_xdc(
+        ts=[numpy.datetime64('2020-08-01'), numpy.datetime64('2020-08-11'), numpy.datetime64('2020-08-21')],
+        bands=['bandzero', 'bandone'], xs=[10., 11., 12., 13., 14.], ys=[20., 21., 22., 23., 24., 25.]
+    )
+    udf_code = _get_udf_code("ndvi01.py")
+    res = execute_local_udf(udf_code, xdc)
+
+    assert isinstance(res, UdfData)
+    result = res.get_datacube_list()[0].get_array()
+
+    assert result.shape == (3, 1, 5, 6)
+    expected = xarray.DataArray(
+        [
+            [_ndvi(0, 100), _ndvi(1, 101)],
+            [_ndvi(10, 110), _ndvi(11, 111)]
+        ],
+        dims=["x", "y"],
+        coords={"t": numpy.datetime64("2020-08-01"), "bands": "ndvi"}
+    )
+    xarray.testing.assert_equal(result[0, 0, 0:2, 0:2], expected)
+
+    assert result[2, 0, 3, 4] == _ndvi(2034, 2134)
+
+
+def test_run_local_udf_from_file_json(tmp_path):
+    udf_code = _get_udf_code("ndvi01.py")
+    xdc = _build_xdc(
+        ts=[numpy.datetime64('2020-08-01'), numpy.datetime64('2020-08-11'), numpy.datetime64('2020-08-21')],
+        bands=['bandzero', 'bandone'], xs=[10., 11., 12., 13., 14.], ys=[20., 21., 22., 23., 24., 25.]
+    )
+    data_path = as_path(tmp_path / "data.json")
+    datacube_to_file(xdc, data_path, fmt="json")
+
+    res = execute_local_udf(udf_code, data_path, fmt="json")
+
+    assert isinstance(res, UdfData)
+    result = res.get_datacube_list()[0].get_array()
+
+    assert result.shape == (3, 1, 5, 6)
+    expected = xarray.DataArray(
+        [
+            [_ndvi(0, 100), _ndvi(1, 101)],
+            [_ndvi(10, 110), _ndvi(11, 111)]
+        ],
+        dims=["x", "y"],
+        coords={"t": numpy.datetime64("2020-08-01"), "bands": "ndvi"}
+    )
+    xarray.testing.assert_equal(result[0, 0, 0:2, 0:2], expected)
+
+    assert result[2, 0, 3, 4] == _ndvi(2034, 2134)
+
+
+def test_run_local_udf_from_file_netcdf(tmp_path):
+    udf_code = _get_udf_code("ndvi01.py")
+    xdc = _build_xdc(
+        ts=[numpy.datetime64('2020-08-01'), numpy.datetime64('2020-08-11'), numpy.datetime64('2020-08-21')],
+        bands=['bandzero', 'bandone'], xs=[10., 11., 12., 13., 14.], ys=[20., 21., 22., 23., 24., 25.]
+    )
+    data_path = as_path(tmp_path / "data.nc")
+    datacube_to_file(xdc, data_path, fmt="netcdf")
+
+    res = execute_local_udf(udf_code, data_path, fmt="netcdf")
+
+    assert isinstance(res, UdfData)
+    result = res.get_datacube_list()[0].get_array()
+
+    assert result.shape == (3, 1, 5, 6)
+    expected = xarray.DataArray(
+        [
+            [_ndvi(0, 100), _ndvi(1, 101)],
+            [_ndvi(10, 110), _ndvi(11, 111)]
+        ],
+        dims=["x", "y"],
+        coords={"t": numpy.datetime64("2020-08-01"), "bands": "ndvi"}
+    )
+    xarray.testing.assert_equal(result[0, 0, 0:2, 0:2], expected)
+
+    assert result[2, 0, 3, 4] == _ndvi(2034, 2134)
